requirements.txt
setuptools
grpcio==1.59.0
grpcio-tools==1.59.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
protobuf==4.24.4